Feature: added alarm screen create modal

+++ app/src/main/java/com/nithiann/alarm/domain/usecase/AddAlarm.kt
@@ -4,7 +4,7 @@ import com.nithiann.alarm.domain.model.Alarm
 import com.nithiann.alarm.domain.repository.AlarmRepository
 
 class AddAlarm(private val alarmRepository: AlarmRepository) {
-    suspend operator fun invoke(alarm: Alarm): Long {
+    suspend operator fun invoke(alarm: Alarm): Int {
         return alarmRepository.addAlarm(alarm);
     }
