+++ app/src/main/java/com/nithiann/alarm/domain/usecase/DeleteAlarm.kt
@@ -3,7 +3,7 @@ package com.nithiann.alarm.domain.usecase
 import com.nithiann.alarm.domain.repository.AlarmRepository
 
 class DeleteAlarm(private val alarmRepository: AlarmRepository) {
-    suspend operator fun invoke(id: Long) {
+    suspend operator fun invoke(id: Int) {
         alarmRepository.deleteAlarm(id)
     }
 }
